requirements.txt
fastapi
uvicorn[standard]
sqlmodel
httpx
apscheduler
python-dotenv
